requirements.txt
Flask==3.0.2
pymongo==4.6.1
twilio==8.13.0
python-dotenv==1.0.0
dnspython==2.7.0
gunicorn==23.0.0
redis==6.2.0